requirements.txt
Package          Version
---------------- -------
aiohappyeyeballs 2.4.3
aiohttp          3.10.8
aiosignal        1.3.1
attrs            24.2.0
discord.py       2.4.0
frozenlist       1.4.1
idna             3.10
multidict        6.1.0
pip              24.2
yarl             1.13.1
